requirements.txt
httpx==0.27.0
fastapi==0.110.0
uvicorn==0.29.0
openai==1.30.1
sentence-transformers==2.7.0
faiss-cpu==1.7.4
scikit-learn==1.4.2
spacy==3.7.4
transformers==4.41.1
torch>=1.12.1
python-dotenv==1.0.1
